Day13 電商產品圖合成器/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
onnxruntime==1.18.1
openai>=1.40.0
tenacity==8.5.0
